Move ImportOptions to Maintenance namespace
